requirements.txt
Django==5.2
gspread==6.2.1
ninja==1.11.1.4
Pillow==11.2.1
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
gunicorn==23.0.0 
whitenoise==6.9.0
django-cors-headers==4.7.0
